factor out common graph stuff
